Introduced linux-source
To avoid having to maintain references to the linux kernel sources in
multiple locations, a native/linux-source build is introduced which
provides the linux kernel source, including configuration and patches.

This build is used by native/crosstool-ng to include the correct linux
userspace API/header files so that it can build a toolchain which
matches the kernel.

Finally, the same linux-source build is used by the linux build to build
the Linux kernel.