Fix qtwebkit crashing in llint_program_prologue()
By default QT uses the gold linker (the parallel GNU linker) if
available. However, this makes qtwebkit crash in
llint_program_prologue() when loading javascript on ARMv7 systems with
hardfp enabled.

This issue is reported to the QT community and has been rasied to a P1
critical priority: https://bugreports.qt.io/browse/QTBUG-44700

Until a proper fix becomes available, we simply disable the use of the
gold linker to bypass the issue.