crosstool-ng: Remove dependency on linux-source
Buildgear v0.9.24+ now rebuilds any build which has any updated
dependencies.

While this is considered correct behaviour it is also very impractical
when making changes to the Linux source build (linux-source) because it
triggers a rebuild of the crosstool-ng toolchain as it depends on
linux-source for Linux kernel-userspace API headers. And rebuilding the
toolchain eventually forces a rebuild of everything... a very taxing
procedure.

To avoid this we simply make crosstool-ng _not_ depend on linux-source
by reconfiguring the crosstool-ng build so that it itself automatically
downloads and installs the Linux kernel-userspace API headers which
matches (API compatible) the kernel version used in linux-source.